prod/dev env

+++ src/events/events.controller.ts
@@ -1,4 +1,4 @@
-import { Body, Controller, Delete, Get, HttpCode, Param, Patch, Post, ValidationPipe } from "@nestjs/common";
+import { Body, Controller, Delete, Get, HttpCode, Logger, NotFoundException, Param, Patch, Post } from "@nestjs/common";
 import { CreateEventDto } from "./dto/create-event.dto";
 import { UpdateEventDto } from "./dto/update-event.dto";
 import { InjectRepository } from "@nestjs/typeorm";
@@ -7,6 +7,8 @@ import { Repository } from "typeorm";
 
 @Controller('/events')
 export class EventsController {
+    private readonly logger = new Logger(EventsController.name)
+
     constructor(
         @InjectRepository(Event)
         private readonly repository: Repository<Event>
@@ -14,12 +16,19 @@ export class EventsController {
 
     @Get()
     async findAll() {
-        return await this.repository.find()
+        this.logger.log('Hit the find route');
+        const events = await this.repository.find();
+        this.logger.debug(`Found ${events.length} events`);
+        return events;
     }
 
     @Get(':id')
     async findOne(@Param('id') id) {
-        return await this.repository.findOneBy({ id: id });
+        const event = await this.repository.findOneBy({ id: id });
+        if (!event) {
+            throw new NotFoundException()
+        };
+        return event;
     }
 
     @Post()
@@ -33,6 +42,9 @@ export class EventsController {
     @Patch(':id')
     async update(@Param('id') id, @Body() input: UpdateEventDto) {
         const event = await this.repository.findOneBy({ id: id });
+        if (!event) {
+            throw new NotFoundException()
+        };
         return await this.repository.save({
             ...event,
             ...input,
@@ -44,6 +56,9 @@ export class EventsController {
     @HttpCode(204)
     async delete(@Param('id') id) {
         const event = await this.repository.findOneBy({ id: id });
+        if (!event) {
+            throw new NotFoundException()
+        };
         await this.repository.remove(event);
     }
 }
